vcmi: move limiters outside of HeroBonus.cpp
This will help for recompilation.
