+++ lib/CHeroHandler.cpp
@@ -24,6 +24,7 @@
 #include "CSkillHandler.h"
 #include "mapObjects/CObjectClassesHandler.h"
 #include "BattleFieldHandler.h"
+#include "bonuses/ILimiter.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
