+++ lib/JsonNode.cpp
@@ -14,6 +14,7 @@
 #include "ScopeGuard.h"
 
 #include "bonuses/HeroBonus.h"
+#include "bonuses/ILimiter.h"
 #include "filesystem/Filesystem.h"
 #include "VCMI_Lib.h" //for identifier resolution
 #include "CModHandler.h"
